requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
apache-beam==2.52.0
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudpickle==2.2.1
crcmod==1.7
dateparser==1.2.0
dill==0.3.1.1
dnspython==2.4.2
docopt==0.6.2
fastavro==1.9.0
fasteners==0.19
frozenlist==1.4.0
google-api-core==2.14.0
google-apitools==0.5.31
google-auth==2.24.0
google-auth-httplib2==0.1.1
google-cloud-aiplatform==1.36.4
google-cloud-bigquery==3.13.0
google-cloud-bigquery-storage==2.23.0
google-cloud-bigtable==2.21.0
google-cloud-core==2.3.3
google-cloud-datastore==2.18.0
google-cloud-dlp==3.13.0
google-cloud-language==2.11.1
google-cloud-pubsub==2.18.4
google-cloud-pubsublite==1.8.3
google-cloud-recommendations-ai==0.10.5
google-cloud-resource-manager==1.10.4
google-cloud-spanner==3.40.1
google-cloud-storage==2.13.0
google-cloud-videointelligence==2.11.4
google-cloud-vision==3.4.5
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.7
grpcio==1.59.3
grpcio-status==1.59.3
hdfs==2.7.3
httplib2==0.22.0
idna==3.6
Js2Py==0.74
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
multidict==6.0.4
numpy==1.24.4
oauth2client==4.1.3
objsize==0.6.1
orjson==3.9.10
overrides==6.5.0
packaging==23.2
proto-plus==1.22.3
protobuf==4.25.1
pyarrow==11.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodome==3.19.0
pydot==1.4.2
pyjsparser==2.7.1
pymongo==4.6.1
pyparsing==3.1.1
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
rpds-py==0.13.2
rsa==4.9
shapely==2.0.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzlocal==5.2
ujson==5.8.0
urllib3==2.1.0
websockets==12.0
yarl==1.9.3
zstandard==0.22.0
